Add optional most_recent_first parameter to get_videos_info method
Added an optional boolean parameter called most_recent_first to the get_videos_info method of the Channel class. This parameter allows the user to specify whether the videos should be sorted by most recent first or by relevance. If most_recent_first is True, the videos will be sorted by date, otherwise they will be sorted by relevance. The default value for most_recent_first is False, so videos are sorted by relevance by default. The parameter has been defined as Optional[bool], so it can be set to None as well.